Restore `cargo fmt` ability to format integration tests
