+++ rustls/Cargo.toml
@@ -10,6 +10,7 @@ homepage = "https://github.com/rustls/rustls"
 repository = "https://github.com/rustls/rustls"
 categories = ["network-programming", "cryptography"]
 autobenches = false
+autotests = false
 build = "build.rs"
 
 [build-dependencies]
@@ -78,6 +79,42 @@ required-features = ["ring"]
 [[example]]
 name = "test_ca"
 path = "examples/internal/test_ca.rs"
+
+[[test]]
+name = "api"
+path = "tests/runners/api.rs"
+
+[[test]]
+name = "api_ffdhe"
+path = "tests/runners/api_ffdhe.rs"
+
+[[test]]
+name = "bogo"
+path = "tests/bogo.rs"
+
+[[test]]
+name = "client_cert_verifier"
+path = "tests/runners/client_cert_verifier.rs"
+
+[[test]]
+name = "ech"
+path = "tests/ech.rs"
+
+[[test]]
+name = "key_log_file_env"
+path = "tests/runners/key_log_file_env.rs"
+
+[[test]]
+name = "process_provider"
+path = "tests/process_provider.rs"
+
+[[test]]
+name = "server_cert_verifier"
+path = "tests/runners/server_cert_verifier.rs"
+
+[[test]]
+name = "unbuffered"
+path = "tests/runners/unbuffered.rs"
 
 [package.metadata.docs.rs]
 # all non-default features except fips (cannot build on docs.rs environment)
