+++ rustls/tests/api.rs
@@ -1,14 +1,7 @@
-#![cfg_attr(read_buf, feature(read_buf))]
-#![cfg_attr(read_buf, feature(core_io_borrowed_buf))]
-
 //! Assorted public API tests.
 
-use std::cell::RefCell;
+#![allow(clippy::duplicate_mod)]
 
-#[macro_use]
-mod macros;
-
-test_for_each_provider! {
 use std::fmt::Debug;
 use std::io::{self, IoSlice, Read, Write};
 use std::ops::{Deref, DerefMut};
@@ -7105,76 +7098,4 @@ fn test_illegal_client_renegotiation_attempt_during_tls12_handshake() {
             got_type: HandshakeType::ClientHello
         }
     );
-}
-
-} // test_for_each_provider!
-
-#[derive(Default, Debug)]
-struct LogCounts {
-    trace: usize,
-    debug: usize,
-    info: usize,
-    warn: usize,
-    error: usize,
-}
-
-impl LogCounts {
-    fn new() -> Self {
-        Self {
-            ..Default::default()
-        }
-    }
-
-    fn reset(&mut self) {
-        *self = Self::new();
-    }
-
-    fn add(&mut self, level: log::Level) {
-        match level {
-            log::Level::Trace => self.trace += 1,
-            log::Level::Debug => self.debug += 1,
-            log::Level::Info => self.info += 1,
-            log::Level::Warn => self.warn += 1,
-            log::Level::Error => self.error += 1,
-        }
-    }
-}
-
-// this must be outside test_for_each_provider!, as we want
-// one thread_local!, not one per provider.
-thread_local!(static COUNTS: RefCell<LogCounts> = RefCell::new(LogCounts::new()));
-
-struct CountingLogger;
-
-#[allow(dead_code)]
-static LOGGER: CountingLogger = CountingLogger;
-
-#[allow(dead_code)]
-impl CountingLogger {
-    fn install() {
-        let _ = log::set_logger(&LOGGER);
-        log::set_max_level(log::LevelFilter::Trace);
-    }
-
-    fn reset() {
-        COUNTS.with(|c| {
-            c.borrow_mut().reset();
-        });
-    }
-}
-
-impl log::Log for CountingLogger {
-    fn enabled(&self, _metadata: &log::Metadata) -> bool {
-        true
-    }
-
-    fn log(&self, record: &log::Record) {
-        println!("logging at {:?}: {:?}", record.level(), record.args());
-
-        COUNTS.with(|c| {
-            c.borrow_mut().add(record.level());
-        });
-    }
-
-    fn flush(&self) {}
 }
