+++ rustls/tests/runners/api_ffdhe.rs
@@ -1,0 +1,7 @@
+#[macro_use]
+mod macros;
+
+test_for_each_provider! {
+    #[path = "../api_ffdhe.rs"]
+    mod tests;
+}
