+++ rustls/tests/runners/unbuffered.rs
@@ -1,0 +1,7 @@
+#[macro_use]
+mod macros;
+
+test_for_each_provider! {
+    #[path = "../unbuffered.rs"]
+    mod tests;
+}
